Creates an extensible, organized, encapsulated query file structure
Now queries and types can be organized by the model (or any common theme) in which they fall under.  And each query resides in its own file.

The queries for a given model (or theme) role up into a manifest of queries.  This identifies all of the queries currently available for the given model (or theme).

Each model’s manifest role’s up into the root query, that is passed to the schema.  Which is the expectation and pattern set by GraphQL and the GraphQL gem.