Implements namespacing at the the schema level
Using folders/directories to namespace GraphQL types does nothing with regards to the actual types that are created at the graphql schema level.

No two types can be identically named, so to enforce namespacing of types, the class name itself has to have a prefix with the namespace in it.

And for the folder structure to be scalable, the namespacing has to exist there, as well.